// ecommerce-client/src/app/features/shipments/models/shipment-status.enum.ts
export enum ShipmentStatus {
    Pending = 'Pending',
    Assigned = 'Assigned',
    LabelCreated = 'LabelCreated',
    PickedUp = 'PickedUp',
    InTransit = 'InTransit',
    OutForDelivery = 'OutForDelivery',
    Delivered = 'Delivered',
    Delayed = 'Delayed',
    Failed = 'Failed',
    Returned = 'Returned',
    Cancelled = 'Cancelled'
}